Restructuring backend apis
