+++ apis/go.mod
@@ -11,6 +11,8 @@ require (
 )
 
 require (
+	github.com/JesusIslam/tldr v0.6.0 // indirect
+	github.com/alixaxel/pagerank v0.0.0-20160306110729-14bfb4c1d88c // indirect
 	github.com/bytedance/sonic v1.12.6 // indirect
 	github.com/bytedance/sonic/loader v0.2.1 // indirect
 	github.com/cloudwego/base64x v0.1.4 // indirect
