+++ apis/user/user.go
@@ -8,18 +8,23 @@ type UserPayload struct {
 	Email     string `json:"email"`
 }
 
-
 type LoginPayload struct {
-	UserName string `json:"username"`
+	Username string `json:"username"`
 	Password string `json:"password"`
 }
 
-
 type User struct {
-	UserId    int64  `json:"userId"`
-	FirstName string `json:"firstName"`
-	LastName  string `json:"lastName"`
-	UserName  string `json:"username"`
-	Password  string `json:"password"`
-	Email     string `json:"email"`
+	UserName  string `json:"username" bson:"username"`
+	Password  string `json:"password" bson:"password"`
+	FirstName string `json:"firstName" bson:"firstName"`
+	LastName  string `json:"lastName" bson:"lastName"`
 }
+
+// type User struct {
+// 	UserId    int64  `json:"userId"`
+// 	FirstName string `json:"firstName"`
+// 	LastName  string `json:"lastName"`
+// 	UserName  string `json:"username"`
+// 	Password  string `json:"password"`
+// 	Email     string `json:"email"`
+// }
