+++ client/src/components/Header/header.js
@@ -16,9 +16,8 @@ function Header({ scrollToContact }) {
     });
   
     const handleProfileSave = (updatedProfile) => {
-      setUserProfile(updatedProfile);
 
-      fetch('http://localhost:8080/user', {
+      fetch('http://localhost:8080/api/user/update', {
         method: 'PUT',
         headers: {
           'Content-Type': 'application/json',
